+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Do not show "Visit Website" button if there is no "titleURL" [#392](https://github.com/rokwire/illinois-app/issues/392).
 - Proper check for online status [#394](https://github.com/rokwire/illinois-app/issues/394).
 - Events sorting in Group [#396](https://github.com/rokwire/illinois-app/issues/396).
+- Group events count [#398](https://github.com/rokwire/illinois-app/issues/398).
 
 ## [2.5.1] - 2021-06-29
 ### Added